% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uslides.R
\name{uslides_document}
\alias{uslides_document}
\title{uslides_document}
\usage{
uslides_document(toc = FALSE, incremental = FALSE, fig_width = 9,
  fig_height = 6, fig_crop = TRUE, fig_caption = TRUE, keep_tex = FALSE,
  pandoc_args = NULL, highlight = "tango", latex_engine = "xelatex")
}
\description{
Document template for uslide beamer presentation.
}
