% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html.R
\name{uslides_html}
\alias{uslides_html}
\title{HTML slideshow presentation}
\usage{
uslides_html(incremental = FALSE, fig_width = 7.5, fig_height = 4.75,
  fig_caption = FALSE, self_contained = FALSE, highlight = "default",
  css = NULL, includes = NULL, keep_md = FALSE, lib_dir = NULL,
  pandoc_args = NULL)
}
\arguments{
\item{incremental}{Logical indicating whether to render bullet points incrementally. Defaults to FALSE. To reverse-order, precede item with `>`}

\item{fig_width}{Default figure width, in inches.}

\item{fig_height}{Default figure height, in inches.}

\item{fig_caption}{Logical indicating whether to render figure captions.}

\item{self_contained}{Logical indicating whether to package presentation as single html file.}

\item{highlight}{Highlight style, if "default" then "haddock" is used.}

\item{css}{Optional, additional css file(s) to include}

\item{includes}{Optional, additional content to include.}

\item{keep_md}{Logical indicating whether to keep .md version of file.}

\item{lib_dir}{If not self contained, name for directory containing presentation files. If NULL (default) then "_files" is added to document name.}

\item{pandoc_args}{Optional, additional args to pass on to pandoc.}
}
\description{
Renders rmarkdown as an HTML slideshow.
}
