% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf.R
\name{uslides_pdf}
\alias{uslides_pdf}
\title{uslides_pdf}
\usage{
uslides_pdf(toc = FALSE, incremental = FALSE, fig_width = 9,
  fig_height = 6, fig_crop = TRUE, fig_caption = TRUE, keep_tex = FALSE,
  pandoc_args = NULL, highlight = "tango", latex_engine = "xelatex")
}
\arguments{
\item{toc}{\code{TRUE} to include a table of contents in the output (only
level 1 headers will be included in the table of contents).}

\item{incremental}{\code{TRUE} to render slide bullets incrementally. Note
that if you want to reverse the default incremental behavior for an
individual bullet you can precede it with \code{>}. For example:
\emph{\code{> - Bullet Text}}}

\item{fig_width}{Default width (in inches) for figures}

\item{fig_height}{Default width (in inches) for figures}

\item{fig_crop}{\code{TRUE} to automatically apply the \code{pdfcrop} utility
(if available) to pdf figures}

\item{fig_caption}{\code{TRUE} to render figures with captions}

\item{keep_tex}{Keep the intermediate tex file used in the conversion to PDF}

\item{pandoc_args}{Additional command line options to pass to pandoc}

\item{highlight}{Syntax highlighting style. Supported styles include
"default", "tango", "pygments", "kate", "monochrome", "espresso",
"zenburn", and "haddock". Pass \code{NULL} to prevent syntax highlighting.}

\item{latex_engine}{LaTeX engine for producing PDF output. Options are
"pdflatex", "lualatex", and "xelatex".}
}
\description{
Document template for uslide beamer presentation.
}
