% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf.R
\name{add_logos}
\alias{add_logos}
\title{add_logos}
\usage{
add_logos(doc_prefix = NULL)
}
\arguments{
\item{doc_prefix}{Path to uslides_pdf_prefix. Default (NULL) will
look up where package is located for default document.}
}
\description{
Set logos as environment variables for uslides design
}
